Added Storage and Modification JavaScript Code
To change every pixel in the canvas, the pixels must be stored in an array. Then, all pixels need to be modified according to their color. To do this, I created a for loop. I needed help writing the code to modify the data array, so I asked a peer. After iterating through the pixels, we have to update the canvas. To do this, we use the putImageData function and run the imageData through the function for context2, which applies to Canvas 2.